requirements.txt
fastapi==0.115.2
uvicorn==0.28.0
pydantic>=2.9.0,<2.10.0,!=2.10.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pandas==2.2.1
scikit-learn==1.4.1.post1
evidently==0.4.22
python-multipart==0.0.7
jinja2==3.1.6
aiofiles==23.2.1
python-dotenv==1.0.1
httpx==0.27.0
prefect==3.3.1
grafana-client==3.1.0
prometheus-client>=0.20.0
plotly==5.19.0
numpy==1.26.4
pytest==7.4.0
pytest-asyncio==0.21.1
urllib3<2.0.0
